model_template/examples/yolo/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
ultralytics>=8.3.0
pydantic>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
